# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Generator Management System API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /generators:
    get:
      operationId: listGenerators
      tags: [generators]
      summary: List all generator records with optional filters
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: generatorId
          in: query
          required: false
          schema:
            type: string
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
      responses:
        "200":
          description: List of generator records
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GeneratorRecord"

    post:
      operationId: createGenerator
      tags: [generators]
      summary: Create a new generator record
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratorInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratorRecord"
...
  /generators/{id}:
    get:
      operationId: getGenerator
      tags: [generators]
      summary: Get a generator record by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Generator record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratorRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    patch:
      operationId: updateGenerator
      tags: [generators]
      summary: Update a generator record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GeneratorUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GeneratorRecord"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteGenerator
      tags: [generators]
      summary: Delete a generator record
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    GeneratorRecord:
      type: object
      properties:
        id:
          type: integer
        tDate:
          type: string
        generatorId:
          type: string
        status:
          type: string
        rating:
          type: ["string", "null"]
        hours:
          type: ["number", "null"]
        remarks:
          type: ["string", "null"]
        deliveryStatus:
          type: ["string", "null"]
        deliveryTo:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - tDate
        - generatorId
        - status
        - createdAt
        - updatedAt
    GeneratorInput:
      type: object
      properties:
        tDate:
          type: string
        generatorId:
          type: string
        status:
          type: string
        rating:
          type: string
        hours:
          type: number
        remarks:
          type: string
      required:
        - tDate
        - generatorId
        - status
    GeneratorUpdate:
      type: object
      properties:
        tDate:
          type: string
        generatorId:
          type: string
        status:
          type: string
        rating:
          type: string
        hours:
          type: number
        remarks:
          type: string
        deliveryStatus:
          type: string
        deliveryTo:
          type: ["string", "null"]
...