# lib/api-zod/src/generated/types/generatorInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Generator Management System API
 * OpenAPI spec version: 0.1.0
 */

export interface GeneratorInput {
  tDate: string;
  generatorId: string;
  status: string;
  rating?: string;
  hours?: number;
  remarks?: string;
}